Change subId from Long to int
Change subId from Long to int

+++ src/com/android/services/telephony/TelephonyConnection.java
@@ -29,7 +29,6 @@ import android.telecom.ConferenceParticipant;
 import android.telecom.CallProperties;
 import android.telecom.Connection;
 import android.telecom.PhoneAccount;
-//FIXME MR1_INTERNAL remove below line after IMS capabilities are moved
 import android.telecom.PhoneCapabilities;
 import android.telecom.PhoneAccountHandle;
 import android.telecom.TelecomManager;
@@ -1132,7 +1131,7 @@ abstract class TelephonyConnection extends Connection {
             }
         }
 
-        long subId = SubscriptionController.getInstance().getSubIdUsingPhoneId(PhoneIdToCall);
+        int subId = SubscriptionController.getInstance().getSubIdUsingPhoneId(PhoneIdToCall);
         if (PhoneIdToCall == SubscriptionManager.INVALID_PHONE_INDEX) {
             Log.d(this,"EMERGENCY_PERM_FAILURE received on all subs, abort redial");
             setDisconnected(DisconnectCauseUtil.toTelecomDisconnectCause(
@@ -1148,7 +1147,7 @@ abstract class TelephonyConnection extends Connection {
                         telecommMgr.getCallCapablePhoneAccounts();
                 for (PhoneAccountHandle handle : phoneAccountHandles) {
                     String sub = handle.getId();
-                    if (Long.toString(subId).equals(sub)){
+                    if (Integer.toString(subId).equals(sub)){
                         Log.d(this,"EMERGENCY REDIAL");
                         for (TelephonyConnectionListener l : mTelephonyListeners) {
                             l.onEmergencyRedial(this, handle, PhoneIdToCall);
